feat: Star 数据页面优化
